pyproject.toml
[project]
name = "research-mcp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.2.0",
    "mcp>=1.12.2",
    "nest-asyncio>=1.6.0",
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
]
